+++ src/envsh/__init__.py
@@ -3,7 +3,6 @@ from .main import (
     read_env,
 )
 
-
 __all__ = [
     'load',
     'read_env',